requirements.txt
numpy>=1.25.2
tweepy>=4.14.0
python-dotenv
tiktoken
emoji>=2.8.0
openai>=0.27.8
pandas>=2.0.3
schedule
requests>=2.31.0
setuptools
APScheduler>=3.10.3
selenium>=4.11.2
python-dateutil
Jinja2
graphviz
discord>=2.3.2
python-telegram-bot
pillow==9.5.0
croniter>=1.4.1
pyyaml